@@ -7,9 +7,14 @@
 # --------------------------------------------------------------------------
 
 from ._batch_management_client import BatchManagementClient
-__all__ = ['BatchManagementClient']
 
-# `._patch.py` is used for handwritten extensions to the generated code
-# Example: https://github.com/Azure/azure-sdk-for-python/blob/main/doc/dev/customize_code/how-to-patch-sdk-code.md
-from ._patch import patch_sdk
-patch_sdk()
+try:
+    from ._patch import __all__ as _patch_all
+    from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+except ImportError:
+    _patch_all = []
+from ._patch import patch_sdk as _patch_sdk
+__all__ = ['BatchManagementClient']
+__all__.extend([p for p in _patch_all if p not in __all__])
+
+_patch_sdk()
