@@ -145,7 +145,9 @@ from ._batch_management_client_enums import (
     ResourceIdentityType,
     StorageAccountType,
 )
-
+from ._patch import __all__ as _patch_all
+from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+from ._patch import patch_sdk as _patch_sdk
 __all__ = [
     'ActivateApplicationPackageParameters',
     'Application',
@@ -283,3 +285,5 @@ __all__ = [
     'ResourceIdentityType',
     'StorageAccountType',
 ]
+__all__.extend([p for p in _patch_all if p not in __all__])
+_patch_sdk()
